LPS-182767 Extract FDSDataProvider methods in CommerceProductDisplayPageTableFDSView to a new CommerceProductDisplayPageFDSDataProvider component
